LICENSE
The MIT License is a permissive open-source license that allows anyone to freely use, modify, distribute, and even sell copies of this software. It requires that the original copyright notice and license text be included in all copies or substantial portions of the software. The software is provided as is, without any warranty, meaning the author bears no responsibility for issues, damages, or liabilities that may arise from its use.